+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Bug fixes**
 
-* Add custom Infura provider which retries queries when bad responses are recieved, which happens occasionally (`@colony/colony-js-client`)
+* Add custom Infura provider which retries queries when bad responses are received, which happens occasionally (`@colony/colony-js-client`)
 
 ## v1.13.3